pyproject.toml
[tool.poetry]
name = "nonebot_plugin_session_orm"
version = "0.1.0"
description = "session 插件 orm 扩展"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-session-orm"
repository = "https://github.com/noneplugin/nonebot-plugin-session-orm"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = { version = "^2.0.1", extras = ["fastapi"] }
nonebot-plugin-session = "^0.2.0b1"
nonebot-plugin-orm = "^0.1.0"

[tool.nonebot]
plugins = ["nonebot_plugin_session_orm"]

[tool.poetry.group.dev.dependencies]
nb-cli = "^1.0.0"
nonebot-plugin-orm = { extras = ["default"], version = "^0.1.0" }

[tool.poetry.group.test.dependencies]
nonebug = "^0.3.0"
pytest-cov = "^4.0.0"
pytest-asyncio = ">=0.20.2,<0.22.0"
asyncpg = "^0.27.0"
aiomysql = "^0.1.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
